refactor(step2):  Frame 중복제거
Frame 구현체의 중복코딩 제거를 위해 extends 구조로 변경
method 이름을 명확히
